Backup is checking if file exist before trying to copy it #7

+++ src/mojeto/cli/backup.py
@@ -1,4 +1,4 @@
-from pathlib import PurePath
+from pathlib import Path, PurePath
 import shutil
 
 from mojeto.config import Config
@@ -12,5 +12,8 @@ class Backup:
     def __call__(self):
         files_to_backup = self.config.get_all_files()
         for file, path in files_to_backup.items():
-            file_path = str(PurePath(path, file))
-            shutil.copy(src=file_path, dst=self.config.repo_location)
+            file_path = Path(PurePath(path, file))
+            if file_path.is_file():
+                shutil.copy(src=file_path, dst=self.config.repo_location)
+            else:
+                print(f"File {file_path} does not exist. Omitting")
